+++ src/shared/components/shadcn/ui/calendar.tsx
@@ -7,8 +7,8 @@ import {
 import { DayButton, DayPicker, getDefaultClassNames } from 'react-day-picker'
 
 import { cn } from '@/lib/utils'
-import { Button } from '@/components/ui/button'
-import { buttonVariants } from '@/components/ui/buttonVariants'
+import { Button } from '@/shared/components/shadcn/ui/button'
+import { buttonVariants } from '@/shared/components/shadcn/ui/buttonVariants'
 
 function Calendar({
   className,
